backend/requirements.txt
Pillow>=9.0.0
flask>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0